requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
alembic==1.11.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
bentoml==1.0.23
black==23.1.0
blessed==1.20.0
blinker==1.6.2
build==0.10.0
cachetools==5.3.0
cattrs==23.1.2
certifi==2022.12.7
charset-normalizer==3.0.1
circus==0.18.0
click==8.1.3
click-option-group==0.5.6
cloudpickle==2.2.1
colorful==0.5.5
contextlib2==21.6.0
contourpy==1.0.7
croniter==1.3.15
cycler==0.11.0
databricks-cli==0.17.7
dateutils==0.6.12
deepdiff==6.3.0
deepmerge==1.1.0
Deprecated==1.2.14
distlib==0.3.6
distro==1.8.0
docker==6.1.3
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.0
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.88.0
filelock==3.12.0
filetype==1.2.0
fire==0.5.0
Flask==2.3.2
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.5.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.16.3
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.59.1
gpustat==1.1
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
hydra-core==1.3.2
hyperopt==0.2.7
idna==3.4
importlib-metadata==6.0.1
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonargparse==4.21.2
jsonschema==4.17.3
kiwisolver==1.4.4
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
line-profiler==4.0.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mlflow==2.4.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.2
nvidia-ml-py==11.525.131
oauthlib==3.2.2
omegaconf==2.3.0
opencensus==0.11.2
opencensus-context==0.1.3
opentelemetry-api==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-aiohttp-client==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
ordered-set==4.1.0
packaging==23.0
pandas==2.0.2
pathspec==0.11.0
pathtools==0.1.2
Pillow==9.4.0
pip-requirements-parser==32.0.1
pip-tools==6.14.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.15.0
pluggy==1.0.0
prometheus-client==0.17.0
protobuf==3.20.3
psutil==5.9.4
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
querystring-parser==1.2.4
ray==2.5.0
readchar==4.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
ruyaml==0.91.0
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.25.1
setproctitle==1.3.2
simple-di==0.1.5
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.22.0
starsessions==1.3.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
torchmetrics==0.11.3
torchview==0.2.6
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typeshed-client==2.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.22.0
virtualenv==20.21.0
wandb==0.15.4
watchfiles==0.19.0
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.5
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
